backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
python-whois==0.8.0
gpt4all==2.7.0
pydantic==2.8.2
tldextract==5.1.2
language-tool-python==2.7.1
google-generativeai
